+++ CHANGELOG.md
@@ -9,6 +9,7 @@
  - Introduce ErrorDetailer interface (#453)
  - module/apmhttp: add CloseIdleConnectons and CancelRequest to RoundTripper (#457)
  - Allow specifying transaction (span) ID via TransactionOptions/SpanOptions (#463)
+ - module/apmzerolog: introduce zerolog log correlation and exception-tracking writer (#428)
 
 ## [v1.2.0](https://github.com/elastic/apm-agent-go/releases/tag/v1.2.0)